creating student stave
