+++ src/views/questions/Question.js
@@ -5,7 +5,7 @@ import styled from 'styled-components'
 import { connect } from 'react-redux'
 import { gradeQuestion } from 'modules/grading/actions'
 import { selectQuestionGrade } from 'modules/grading/selectors'
-import Stave from 'views/music/Stave'
+import StudentStave from 'views/music/StudentStave'
 import type { Question as QuestionType } from 'modules/student-test'
 import type { QuestionGrade } from 'modules/grading'
 import type { Dispatch } from 'redux'
@@ -49,7 +49,7 @@ class Question extends Component {
           {this.props.question.statement}
         </div>
         <div className="question-notation">
-          <Stave notes={this.props.question.notation} question={this.props.question}/>
+          <StudentStave notes={this.props.question.notation} question={this.props.question}/>
         </div>
       </div>
     )
